+++ app/src/main/java/com/zeapo/pwdstore/utils/PasswordRepository.kt
@@ -4,6 +4,7 @@
  */
 package com.zeapo.pwdstore.utils
 
+import android.content.Context
 import android.content.SharedPreferences
 import androidx.core.content.edit
 import com.zeapo.pwdstore.Application
@@ -29,6 +30,18 @@ open class PasswordRepository protected constructor() {
 
         INDEPENDENT(Comparator { p1: PasswordItem, p2: PasswordItem ->
             p1.name.compareTo(p2.name, ignoreCase = true)
+        }),
+
+        RECENTLY_USED(Comparator { p1: PasswordItem, p2: PasswordItem ->
+            val recentHistory = Application.instance.getSharedPreferences("recent_password_history", Context.MODE_PRIVATE)
+            val timeP1 = recentHistory.getString(p1.file.absolutePath.base64())
+            val timeP2 = recentHistory.getString(p2.file.absolutePath.base64())
+            when {
+                timeP1 != null && timeP2 != null -> timeP2.compareTo(timeP1)
+                timeP1 != null && timeP2 == null -> return@Comparator -1
+                timeP1 == null && timeP2 != null -> return@Comparator 1
+                else -> p1.name.compareTo(p2.name, ignoreCase = true)
+            }
         }),
 
         FILE_FIRST(Comparator { p1: PasswordItem, p2: PasswordItem ->